NDRS-186: Apply remaining review comments.

+++ types/src/auction/types.rs
@@ -1,2 +1,2 @@
-/// Representation of delegation rate of tokens.
-pub type DelegationRate = u32;
+/// Representation of delegation rate as a tiny fraction of tokens.
+pub type DelegationRate = u64;
